requirements.txt
# Core dependencies for TFT Voice Assistant
# Updated to latest secure versions as of 2025-01-19

# Web scraping and HTML parsing
beautifulsoup4==4.13.4
requests==2.32.5
lxml==6.0.0

# OCR and image processing
opencv-python==4.11.0.86
pillow==11.3.0
pytesseract==0.3.13

# Screen automation
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
MouseInfo==0.1.3

# Voice recognition and TTS
SpeechRecognition==3.10.4
pyttsx3==2.90
PyAudio==0.2.14

# Fuzzy string matching
RapidFuzz==3.13.0

# Data processing
numpy==2.2.6
pandas==2.3.0

# Browser automation (for advanced scraping if needed)
selenium==4.35.0

# Configuration and environment
python-dotenv==1.0.1

# Input handling
pynput==1.7.7

# HTTP and networking
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# Date/time utilities
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Async support
trio==0.30.0
trio-websocket==0.12.2
sniffio==1.3.1
outcome==1.3.0.post0

# WebSocket support
websocket-client==1.8.0
wsproto==1.2.0

# Networking protocols
h11==0.16.0
PySocks==1.7.1

# Utilities
attrs==25.3.0
packaging==25.0
six==1.17.0
sortedcontainers==2.4.0
soupsieve==2.7
typing_extensions==4.14.1
rubicon-objc==0.5.2

# Development and testing dependencies (install with pip install -e .[dev])
# pytest==8.3.3
# pytest-cov==6.0.0
# pytest-asyncio==0.25.0
# black==24.10.0
# flake8==7.1.1
# mypy==1.13.0
# pre-commit==4.0.1
